Flatten story hierarchy for simple stories (#208)
Many components have a single story. Instead of adding a folder with a single story, the story is now hoisted to an item in the containing folder.

Storybook avoids creating the extra folder as long as there is only one named export from the component module, and the name matches the content path for the story.